+++ applications/calendar/CHANGELOG.md
@@ -7,7 +7,7 @@
 
 # [4.0.0-beta.12] - 2021-01-21
 ## Added
-- Send invitations from ProtonCalendar
+- Send invitations from Proton Calendar
 
 ## Updated
 - Harmonise answer for recurring events with single edits for all other calendar providers
@@ -163,7 +163,7 @@
 
 # Beta launch [4.0.0-beta.0] - 2019-12-30
 
-Release of ProtonCalendar on Beta https://protonmail.com/blog/protoncalendar-beta-announcement/
+Release of Proton Calendar on Beta https://protonmail.com/blog/protoncalendar-beta-announcement/
 
 ## Added